refactor: location of connecting prices to txs methods
